pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videohub-controller"
version = "0.1.0"
description = "Videohub Controller - Blackmagic Videohub 10x10 routing GUI"
authors = [{name = "Chad Littlepage", email = "chad.littlepage@gmail.com"}]
requires-python = ">=3.12"
dependencies = [
    "pyobjc-core>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
]

[project.scripts]
videohub-controller = "videohub_controller.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
